updated git ignore , CMakeLists.txt generate files

+++ scripts/generate_windows.bat
@@ -3,8 +3,14 @@
 :: Ensure script exits on error
 setlocal enabledelayedexpansion
 
+:: Check if VCPKG_ROOT is set
+if "%VCPKG_ROOT%"=="" (
+    echo VCPKG_ROOT is not set. Please set it to the vcpkg installation path.
+    exit /b 1
+)
+
 :: Delete existing build directory if it exists
-rmdir /s /q build
+if exist build rmdir /s /q build
 
 :: Generate build files
 cmake -S . -B build -G "Visual Studio 17 2022" ^
